refactor: `FileUtil.java` remove main method and small param refactor
