+++ at_client/src/main/java/org/atsign/client/util/FileUtil.java
@@ -1,0 +1,26 @@
+package org.atsign.client.util;
+
+import java.io.*;
+public class FileUtil
+{
+	
+	// This method converts a file to a byte array and returns the byte array
+	public static byte[] fileToByteArray(File file)throws IOException
+	{
+		FileInputStream fis = new FileInputStream(file);
+        byte[] data = new byte[(int)file.length()];
+        fis.read(data);
+        fis.close();
+		return data;
+	}
+	
+	// This method converts a byte array to a file and returns the file
+	public static File byteArrayToFile(byte data[], String fileName)throws IOException
+	{
+		File file = new File(fileName);
+		OutputStream os = new FileOutputStream(file);
+        os.write(data);
+        os.close();
+		return file;
+	}
+}
